CMake build
This CMakeLists.txt builds and installs the BON library and the bon2json
and json2bon utilities.

It forces the language level of non-MSVC compilers to C99 as constructs
in the source code rely on C99-like declarations of identifiers. It's
not quite perfect, as not all non-MSVC compilers understand the -std=
directive, an enumeration of Clang/GCC/Intel-Linux/PathScale would
probably be better.